requirements.txt
pillow
fpdf